# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Job": {
        "additionalProperties": false,
        "properties": {
          "allowed_windows": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Allowed Windows",
            "type": "array"
          },
          "asset_id": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Asset Id",
            "type": "string"
          },
          "department": {
            "enum": [
              "TRACK",
              "SIGNAL",
              "ELECTRICAL",
              "CIVIL"
            ],
            "title": "Department",
            "type": "string"
          },
          "duration_max_minutes": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Max Minutes",
            "type": "integer"
          },
          "duration_min_minutes": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Min Minutes",
            "type": "integer"
          },
          "duration_minutes": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "job_id": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Job Id",
            "type": "string"
          },
          "priority": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "required_resources": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Required Resources",
            "type": "array"
          },
          "section_id": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Section Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "work_type": {
            "minLength": 1,
            "title": "Work Type",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "department",
          "asset_id",
          "section_id",
          "work_type",
          "priority",
          "duration_minutes",
          "duration_min_minutes",
          "duration_max_minutes",
          "required_resources",
          "allowed_windows",
          "status"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "UNSCHEDULED",
          "SCHEDULED",
          "LOCKED",
          "REJECTED",
          "INVALID"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "LockRequest": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "reason"
        ],
        "title": "LockRequest",
        "type": "object"
      },
      "LockResponse": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "job_id",
          "locked",
          "reason_codes"
        ],
        "title": "LockResponse",
        "type": "object"
      },
...
      "PlanningRunState": {
        "description": "Terminal states only.\n\nPlanning is synchronous: POST /planning-runs solves before it responds, so a\nrun is never observable in flight. QUEUED and RUNNING existed here but were\nassigned nowhere, which advertised an execution model the system did not\nhave. Reintroduce them together with a real queue, not before - see\ndocs/architecture.md, \"Execution model\".",
        "enum": [
          "FEASIBLE",
          "OPTIMAL",
          "INFEASIBLE",
          "TIMEOUT",
          "INVALID",
          "FAILED"
        ],
        "title": "PlanningRunState",
        "type": "string"
      },
...
      "RapidBlockDetail": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "actor_role": {
            "title": "Actor Role",
            "type": "string"
          },
          "base_run_id": {
            "title": "Base Run Id",
            "type": "string"
          },
          "base_snapshot_id": {
            "title": "Base Snapshot Id",
            "type": "string"
          },
          "candidate_plan_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanningRunState"
              },
              {
                "type": "null"
              }
            ]
          },
          "changed_jobs": {
            "additionalProperties": {
              "enum": [
                "SCHEDULED",
                "REJECTED",
                "PRESERVED",
                "CHANGED"
              ],
              "type": "string"
            },
            "title": "Changed Jobs",
            "type": "object"
          },
          "child_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Run Id"
          },
          "derived_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived Snapshot Id"
          },
          "detail_url": {
            "title": "Detail Url",
            "type": "string"
          },
          "justification": {
            "title": "Justification",
            "type": "string"
          },
          "preserved_locked_jobs": {
            "items": {
              "type": "string"
            },
            "title": "Preserved Locked Jobs",
            "type": "array"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source_reported_at": {
            "format": "date-time",
            "title": "Source Reported At",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RapidBlockState"
          },
          "urgent_job": {
            "$ref": "#/components/schemas/Job"
          },
          "validator": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidatorSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "request_id",
          "state",
          "base_run_id",
          "base_snapshot_id",
          "reason_codes",
          "detail_url",
          "actor",
          "actor_role",
          "justification",
          "source_reported_at",
          "urgent_job",
          "changed_jobs",
          "preserved_locked_jobs"
        ],
        "title": "RapidBlockDetail",
        "type": "object"
      },
      "RapidBlockRequestCreate": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "actor_role": {
            "const": "PLANNER",
            "title": "Actor Role",
            "type": "string"
          },
          "base_run_id": {
            "minLength": 1,
            "title": "Base Run Id",
            "type": "string"
          },
          "justification": {
            "minLength": 1,
            "title": "Justification",
            "type": "string"
          },
          "source_reported_at": {
            "format": "date-time",
            "title": "Source Reported At",
            "type": "string"
          },
          "urgent_job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "required": [
          "base_run_id",
          "actor",
          "actor_role",
          "justification",
          "source_reported_at",
          "urgent_job"
        ],
        "title": "RapidBlockRequestCreate",
        "type": "object"
      },
      "RapidBlockResponse": {
        "additionalProperties": false,
        "properties": {
          "base_run_id": {
            "title": "Base Run Id",
            "type": "string"
          },
          "base_snapshot_id": {
            "title": "Base Snapshot Id",
            "type": "string"
          },
          "child_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Run Id"
          },
          "derived_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived Snapshot Id"
          },
          "detail_url": {
            "title": "Detail Url",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RapidBlockState"
          }
        },
        "required": [
          "request_id",
          "state",
          "base_run_id",
          "base_snapshot_id",
          "reason_codes",
          "detail_url"
        ],
        "title": "RapidBlockResponse",
        "type": "object"
      },
      "RapidBlockState": {
        "enum": [
          "SUBMITTED",
          "VALIDATING",
          "REJECTED",
          "PLANNING",
          "CANDIDATE_READY",
          "NO_CANDIDATE"
        ],
        "title": "RapidBlockState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ValidatorSummary": {
        "additionalProperties": false,
        "properties": {
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Issues",
            "type": "array"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "validated_at": {
            "format": "date-time",
            "title": "Validated At",
            "type": "string"
          }
        },
        "required": [
          "passed",
          "validated_at"
        ],
        "title": "ValidatorSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Decision-support API for the RailNiyojan SIH26027 vertical slice.",
    "title": "RailNiyojan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/planning-runs/{run_id}/lock": {
      "post": {
        "operationId": "lock_schedule_item_planning_runs__run_id__lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Schedule Item",
        "tags": [
          "planning-runs"
        ]
      }
    },
...
    "/railradar/trains/{train_number}/route": {
      "get": {
        "operationId": "route_railradar_trains__train_number__route_get",
        "parameters": [
          {
            "in": "path",
            "name": "train_number",
            "required": true,
            "schema": {
              "title": "Train Number",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "geojson",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stops",
            "required": false,
            "schema": {
              "default": false,
              "title": "Stops",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route Railradar Trains  Train Number  Route Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route",
        "tags": [
          "railradar"
        ]
      }
    },
...
    "/rapidblock-requests": {
      "post": {
        "operationId": "create_rapidblock_request_rapidblock_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RapidBlockRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RapidBlockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rapidblock Request",
        "tags": [
          "rapidblock"
        ]
      }
    },
    "/rapidblock-requests/{request_id}": {
      "get": {
        "operationId": "get_rapidblock_request_rapidblock_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RapidBlockDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rapidblock Request",
        "tags": [
          "rapidblock"
        ]
      }
    }
  }
}